@@ -2,10 +2,10 @@ name: webview_flutter_android
 description: A Flutter plugin that provides a WebView widget on Android.
 repository: https://github.com/flutter/plugins/tree/main/packages/webview_flutter/webview_flutter_android
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+webview%22
-version: 2.10.4
+version: 3.0.0
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
   flutter: ">=3.0.0"
 
 flutter:
@@ -15,11 +15,12 @@ flutter:
       android:
         package: io.flutter.plugins.webviewflutter
         pluginClass: WebViewFlutterPlugin
+        dartPluginClass: AndroidWebViewPlatform
 
 dependencies:
   flutter:
     sdk: flutter
-  webview_flutter_platform_interface: ^1.8.0
+  webview_flutter_platform_interface: ^2.0.0
 
 dev_dependencies:
   build_runner: ^2.1.4
@@ -27,5 +28,5 @@ dev_dependencies:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  mockito: ^5.2.0
-  pigeon: ^4.0.2
+  mockito: ^5.3.2
+  pigeon: ^4.2.3
