@@ -6,7 +6,6 @@ package io.flutter.plugins.webviewflutter;
 
 import android.webkit.DownloadListener;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.DownloadListenerHostApi;
 
 /**
@@ -21,11 +20,9 @@ public class DownloadListenerHostApiImpl implements DownloadListenerHostApi {
 
   /**
    * Implementation of {@link DownloadListener} that passes arguments of callback methods to Dart.
-   *
-   * <p>No messages are sent to Dart after {@link DownloadListenerImpl#release} is called.
    */
-  public static class DownloadListenerImpl implements DownloadListener, Releasable {
-    @Nullable private DownloadListenerFlutterApiImpl flutterApi;
+  public static class DownloadListenerImpl implements DownloadListener {
+    private final DownloadListenerFlutterApiImpl flutterApi;
 
     /**
      * Creates a {@link DownloadListenerImpl} that passes arguments of callbacks methods to Dart.
@@ -43,18 +40,8 @@ public class DownloadListenerHostApiImpl implements DownloadListenerHostApi {
         String contentDisposition,
         String mimetype,
         long contentLength) {
-      if (flutterApi != null) {
-        flutterApi.onDownloadStart(
-            this, url, userAgent, contentDisposition, mimetype, contentLength, reply -> {});
-      }
-    }
-
-    @Override
-    public void release() {
-      if (flutterApi != null) {
-        flutterApi.dispose(this, reply -> {});
-      }
-      flutterApi = null;
+      flutterApi.onDownloadStart(
+          this, url, userAgent, contentDisposition, mimetype, contentLength, reply -> {});
     }
   }
 
