@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v4.0.2), do not edit directly.
+// Autogenerated from Pigeon (v4.2.3), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package io.flutter.plugins.webviewflutter;
@@ -17,6 +17,7 @@ import java.io.ByteArrayOutputStream;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -342,22 +343,22 @@ public class GeneratedAndroidWebView {
 
     void error(Throwable error);
   }
-
-  private static class JavaObjectHostApiCodec extends StandardMessageCodec {
-    public static final JavaObjectHostApiCodec INSTANCE = new JavaObjectHostApiCodec();
-
-    private JavaObjectHostApiCodec() {}
-  }
-
-  /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
+  /**
+   * Handles methods calls to the native Java Object class.
+   *
+   * <p>Also handles calls to remove the reference to an instance with `dispose`.
+   *
+   * <p>See https://docs.oracle.com/javase/7/docs/api/java/lang/Object.html.
+   *
+   * <p>Generated interface from Pigeon that represents a handler of messages from Flutter.
+   */
   public interface JavaObjectHostApi {
     void dispose(@NonNull Long identifier);
 
     /** The codec used by JavaObjectHostApi. */
     static MessageCodec<Object> getCodec() {
-      return JavaObjectHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `JavaObjectHostApi` to handle messages through the `binaryMessenger`.
      */
@@ -372,6 +373,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number identifierArg = (Number) args.get(0);
                   if (identifierArg == null) {
                     throw new NullPointerException("identifierArg unexpectedly null.");
@@ -389,14 +391,13 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class JavaObjectFlutterApiCodec extends StandardMessageCodec {
-    public static final JavaObjectFlutterApiCodec INSTANCE = new JavaObjectFlutterApiCodec();
-
-    private JavaObjectFlutterApiCodec() {}
-  }
-
-  /** Generated class from Pigeon that represents Flutter messages that can be called from Java. */
+  /**
+   * Handles callbacks methods for the native Java Object class.
+   *
+   * <p>See https://docs.oracle.com/javase/7/docs/api/java/lang/Object.html.
+   *
+   * <p>Generated class from Pigeon that represents Flutter messages that can be called from Java.
+   */
   public static class JavaObjectFlutterApi {
     private final BinaryMessenger binaryMessenger;
 
@@ -407,9 +408,9 @@ public class GeneratedAndroidWebView {
     public interface Reply<T> {
       void reply(T reply);
     }
-
+    /** The codec used by JavaObjectFlutterApi. */
     static MessageCodec<Object> getCodec() {
-      return JavaObjectFlutterApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
 
     public void dispose(@NonNull Long identifierArg, Reply<Void> callback) {
@@ -417,19 +418,12 @@ public class GeneratedAndroidWebView {
           new BasicMessageChannel<>(
               binaryMessenger, "dev.flutter.pigeon.JavaObjectFlutterApi.dispose", getCodec());
       channel.send(
-          new ArrayList<Object>(Arrays.asList(identifierArg)),
+          new ArrayList<Object>(Collections.singletonList(identifierArg)),
           channelReply -> {
             callback.reply(null);
           });
     }
   }
-
-  private static class CookieManagerHostApiCodec extends StandardMessageCodec {
-    public static final CookieManagerHostApiCodec INSTANCE = new CookieManagerHostApiCodec();
-
-    private CookieManagerHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface CookieManagerHostApi {
     void clearCookies(Result<Boolean> result);
@@ -438,9 +432,8 @@ public class GeneratedAndroidWebView {
 
     /** The codec used by CookieManagerHostApi. */
     static MessageCodec<Object> getCodec() {
-      return CookieManagerHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `CookieManagerHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -490,6 +483,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   String urlArg = (String) args.get(0);
                   if (urlArg == null) {
                     throw new NullPointerException("urlArg unexpectedly null.");
@@ -518,7 +512,7 @@ public class GeneratedAndroidWebView {
     private WebViewHostApiCodec() {}
 
     @Override
-    protected Object readValueOfType(byte type, ByteBuffer buffer) {
+    protected Object readValueOfType(byte type, @NonNull ByteBuffer buffer) {
       switch (type) {
         case (byte) 128:
           return WebViewPoint.fromMap((Map<String, Object>) readValue(buffer));
@@ -529,7 +523,7 @@ public class GeneratedAndroidWebView {
     }
 
     @Override
-    protected void writeValue(ByteArrayOutputStream stream, Object value) {
+    protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
       if (value instanceof WebViewPoint) {
         stream.write(128);
         writeValue(stream, ((WebViewPoint) value).toMap());
@@ -542,8 +536,6 @@ public class GeneratedAndroidWebView {
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface WebViewHostApi {
     void create(@NonNull Long instanceId, @NonNull Boolean useHybridComposition);
-
-    void dispose(@NonNull Long instanceId);
 
     void loadData(
         @NonNull Long instanceId,
@@ -619,7 +611,6 @@ public class GeneratedAndroidWebView {
     static MessageCodec<Object> getCodec() {
       return WebViewHostApiCodec.INSTANCE;
     }
-
     /** Sets up an instance of `WebViewHostApi` to handle messages through the `binaryMessenger`. */
     static void setup(BinaryMessenger binaryMessenger, WebViewHostApi api) {
       {
@@ -632,6 +623,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -656,31 +648,6 @@ public class GeneratedAndroidWebView {
       {
         BasicMessageChannel<Object> channel =
             new BasicMessageChannel<>(
-                binaryMessenger, "dev.flutter.pigeon.WebViewHostApi.dispose", getCodec());
-        if (api != null) {
-          channel.setMessageHandler(
-              (message, reply) -> {
-                Map<String, Object> wrapped = new HashMap<>();
-                try {
-                  ArrayList<Object> args = (ArrayList<Object>) message;
-                  Number instanceIdArg = (Number) args.get(0);
-                  if (instanceIdArg == null) {
-                    throw new NullPointerException("instanceIdArg unexpectedly null.");
-                  }
-                  api.dispose((instanceIdArg == null) ? null : instanceIdArg.longValue());
-                  wrapped.put("result", null);
-                } catch (Error | RuntimeException exception) {
-                  wrapped.put("error", wrapError(exception));
-                }
-                reply.reply(wrapped);
-              });
-        } else {
-          channel.setMessageHandler(null);
-        }
-      }
-      {
-        BasicMessageChannel<Object> channel =
-            new BasicMessageChannel<>(
                 binaryMessenger, "dev.flutter.pigeon.WebViewHostApi.loadData", getCodec());
         if (api != null) {
           channel.setMessageHandler(
@@ -688,6 +655,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -725,6 +693,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -764,6 +733,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -800,6 +770,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -834,6 +805,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -860,6 +832,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -886,6 +859,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -912,6 +886,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -937,6 +912,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -962,6 +938,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -987,6 +964,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1020,6 +998,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1064,6 +1043,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1090,6 +1070,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1126,6 +1107,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1162,6 +1144,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1188,6 +1171,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1214,6 +1198,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1243,6 +1228,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Boolean enabledArg = (Boolean) args.get(0);
                   if (enabledArg == null) {
                     throw new NullPointerException("enabledArg unexpectedly null.");
@@ -1268,6 +1254,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1303,6 +1290,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1339,6 +1327,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1375,6 +1364,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1405,6 +1395,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1435,6 +1426,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1458,18 +1450,9 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class WebSettingsHostApiCodec extends StandardMessageCodec {
-    public static final WebSettingsHostApiCodec INSTANCE = new WebSettingsHostApiCodec();
-
-    private WebSettingsHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface WebSettingsHostApi {
     void create(@NonNull Long instanceId, @NonNull Long webViewInstanceId);
-
-    void dispose(@NonNull Long instanceId);
 
     void setDomStorageEnabled(@NonNull Long instanceId, @NonNull Boolean flag);
 
@@ -1497,9 +1480,8 @@ public class GeneratedAndroidWebView {
 
     /** The codec used by WebSettingsHostApi. */
     static MessageCodec<Object> getCodec() {
-      return WebSettingsHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `WebSettingsHostApi` to handle messages through the `binaryMessenger`.
      */
@@ -1514,6 +1496,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1538,31 +1521,6 @@ public class GeneratedAndroidWebView {
       {
         BasicMessageChannel<Object> channel =
             new BasicMessageChannel<>(
-                binaryMessenger, "dev.flutter.pigeon.WebSettingsHostApi.dispose", getCodec());
-        if (api != null) {
-          channel.setMessageHandler(
-              (message, reply) -> {
-                Map<String, Object> wrapped = new HashMap<>();
-                try {
-                  ArrayList<Object> args = (ArrayList<Object>) message;
-                  Number instanceIdArg = (Number) args.get(0);
-                  if (instanceIdArg == null) {
-                    throw new NullPointerException("instanceIdArg unexpectedly null.");
-                  }
-                  api.dispose((instanceIdArg == null) ? null : instanceIdArg.longValue());
-                  wrapped.put("result", null);
-                } catch (Error | RuntimeException exception) {
-                  wrapped.put("error", wrapError(exception));
-                }
-                reply.reply(wrapped);
-              });
-        } else {
-          channel.setMessageHandler(null);
-        }
-      }
-      {
-        BasicMessageChannel<Object> channel =
-            new BasicMessageChannel<>(
                 binaryMessenger,
                 "dev.flutter.pigeon.WebSettingsHostApi.setDomStorageEnabled",
                 getCodec());
@@ -1572,6 +1530,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1604,6 +1563,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1636,6 +1596,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1668,6 +1629,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1700,6 +1662,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1730,6 +1693,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1762,6 +1726,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1794,6 +1759,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1826,6 +1792,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1858,6 +1825,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1890,6 +1858,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1922,6 +1891,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1944,23 +1914,14 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class JavaScriptChannelHostApiCodec extends StandardMessageCodec {
-    public static final JavaScriptChannelHostApiCodec INSTANCE =
-        new JavaScriptChannelHostApiCodec();
-
-    private JavaScriptChannelHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface JavaScriptChannelHostApi {
     void create(@NonNull Long instanceId, @NonNull String channelName);
 
     /** The codec used by JavaScriptChannelHostApi. */
     static MessageCodec<Object> getCodec() {
-      return JavaScriptChannelHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `JavaScriptChannelHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -1976,6 +1937,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -1998,14 +1960,6 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class JavaScriptChannelFlutterApiCodec extends StandardMessageCodec {
-    public static final JavaScriptChannelFlutterApiCodec INSTANCE =
-        new JavaScriptChannelFlutterApiCodec();
-
-    private JavaScriptChannelFlutterApiCodec() {}
-  }
-
   /** Generated class from Pigeon that represents Flutter messages that can be called from Java. */
   public static class JavaScriptChannelFlutterApi {
     private final BinaryMessenger binaryMessenger;
@@ -2017,22 +1971,9 @@ public class GeneratedAndroidWebView {
     public interface Reply<T> {
       void reply(T reply);
     }
-
+    /** The codec used by JavaScriptChannelFlutterApi. */
     static MessageCodec<Object> getCodec() {
-      return JavaScriptChannelFlutterApiCodec.INSTANCE;
-    }
-
-    public void dispose(@NonNull Long instanceIdArg, Reply<Void> callback) {
-      BasicMessageChannel<Object> channel =
-          new BasicMessageChannel<>(
-              binaryMessenger,
-              "dev.flutter.pigeon.JavaScriptChannelFlutterApi.dispose",
-              getCodec());
-      channel.send(
-          new ArrayList<Object>(Arrays.asList(instanceIdArg)),
-          channelReply -> {
-            callback.reply(null);
-          });
+      return new StandardMessageCodec();
     }
 
     public void postMessage(
@@ -2049,22 +1990,17 @@ public class GeneratedAndroidWebView {
           });
     }
   }
-
-  private static class WebViewClientHostApiCodec extends StandardMessageCodec {
-    public static final WebViewClientHostApiCodec INSTANCE = new WebViewClientHostApiCodec();
-
-    private WebViewClientHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface WebViewClientHostApi {
-    void create(@NonNull Long instanceId, @NonNull Boolean shouldOverrideUrlLoading);
+    void create(@NonNull Long instanceId);
+
+    void setSynchronousReturnValueForShouldOverrideUrlLoading(
+        @NonNull Long instanceId, @NonNull Boolean value);
 
     /** The codec used by WebViewClientHostApi. */
     static MessageCodec<Object> getCodec() {
-      return WebViewClientHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `WebViewClientHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -2080,18 +2016,45 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
                   }
-                  Boolean shouldOverrideUrlLoadingArg = (Boolean) args.get(1);
-                  if (shouldOverrideUrlLoadingArg == null) {
-                    throw new NullPointerException(
-                        "shouldOverrideUrlLoadingArg unexpectedly null.");
+                  api.create((instanceIdArg == null) ? null : instanceIdArg.longValue());
+                  wrapped.put("result", null);
+                } catch (Error | RuntimeException exception) {
+                  wrapped.put("error", wrapError(exception));
+                }
+                reply.reply(wrapped);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger,
+                "dev.flutter.pigeon.WebViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading",
+                getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                Map<String, Object> wrapped = new HashMap<>();
+                try {
+                  ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
+                  Number instanceIdArg = (Number) args.get(0);
+                  if (instanceIdArg == null) {
+                    throw new NullPointerException("instanceIdArg unexpectedly null.");
                   }
-                  api.create(
-                      (instanceIdArg == null) ? null : instanceIdArg.longValue(),
-                      shouldOverrideUrlLoadingArg);
+                  Boolean valueArg = (Boolean) args.get(1);
+                  if (valueArg == null) {
+                    throw new NullPointerException("valueArg unexpectedly null.");
+                  }
+                  api.setSynchronousReturnValueForShouldOverrideUrlLoading(
+                      (instanceIdArg == null) ? null : instanceIdArg.longValue(), valueArg);
                   wrapped.put("result", null);
                 } catch (Error | RuntimeException exception) {
                   wrapped.put("error", wrapError(exception));
@@ -2111,7 +2074,7 @@ public class GeneratedAndroidWebView {
     private WebViewClientFlutterApiCodec() {}
 
     @Override
-    protected Object readValueOfType(byte type, ByteBuffer buffer) {
+    protected Object readValueOfType(byte type, @NonNull ByteBuffer buffer) {
       switch (type) {
         case (byte) 128:
           return WebResourceErrorData.fromMap((Map<String, Object>) readValue(buffer));
@@ -2125,7 +2088,7 @@ public class GeneratedAndroidWebView {
     }
 
     @Override
-    protected void writeValue(ByteArrayOutputStream stream, Object value) {
+    protected void writeValue(@NonNull ByteArrayOutputStream stream, Object value) {
       if (value instanceof WebResourceErrorData) {
         stream.write(128);
         writeValue(stream, ((WebResourceErrorData) value).toMap());
@@ -2149,20 +2112,9 @@ public class GeneratedAndroidWebView {
     public interface Reply<T> {
       void reply(T reply);
     }
-
+    /** The codec used by WebViewClientFlutterApi. */
     static MessageCodec<Object> getCodec() {
       return WebViewClientFlutterApiCodec.INSTANCE;
-    }
-
-    public void dispose(@NonNull Long instanceIdArg, Reply<Void> callback) {
-      BasicMessageChannel<Object> channel =
-          new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.WebViewClientFlutterApi.dispose", getCodec());
-      channel.send(
-          new ArrayList<Object>(Arrays.asList(instanceIdArg)),
-          channelReply -> {
-            callback.reply(null);
-          });
     }
 
     public void onPageStarted(
@@ -2275,22 +2227,14 @@ public class GeneratedAndroidWebView {
           });
     }
   }
-
-  private static class DownloadListenerHostApiCodec extends StandardMessageCodec {
-    public static final DownloadListenerHostApiCodec INSTANCE = new DownloadListenerHostApiCodec();
-
-    private DownloadListenerHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface DownloadListenerHostApi {
     void create(@NonNull Long instanceId);
 
     /** The codec used by DownloadListenerHostApi. */
     static MessageCodec<Object> getCodec() {
-      return DownloadListenerHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `DownloadListenerHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -2306,6 +2250,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -2323,14 +2268,6 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class DownloadListenerFlutterApiCodec extends StandardMessageCodec {
-    public static final DownloadListenerFlutterApiCodec INSTANCE =
-        new DownloadListenerFlutterApiCodec();
-
-    private DownloadListenerFlutterApiCodec() {}
-  }
-
   /** Generated class from Pigeon that represents Flutter messages that can be called from Java. */
   public static class DownloadListenerFlutterApi {
     private final BinaryMessenger binaryMessenger;
@@ -2342,20 +2279,9 @@ public class GeneratedAndroidWebView {
     public interface Reply<T> {
       void reply(T reply);
     }
-
+    /** The codec used by DownloadListenerFlutterApi. */
     static MessageCodec<Object> getCodec() {
-      return DownloadListenerFlutterApiCodec.INSTANCE;
-    }
-
-    public void dispose(@NonNull Long instanceIdArg, Reply<Void> callback) {
-      BasicMessageChannel<Object> channel =
-          new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.DownloadListenerFlutterApi.dispose", getCodec());
-      channel.send(
-          new ArrayList<Object>(Arrays.asList(instanceIdArg)),
-          channelReply -> {
-            callback.reply(null);
-          });
+      return new StandardMessageCodec();
     }
 
     public void onDownloadStart(
@@ -2385,22 +2311,14 @@ public class GeneratedAndroidWebView {
           });
     }
   }
-
-  private static class WebChromeClientHostApiCodec extends StandardMessageCodec {
-    public static final WebChromeClientHostApiCodec INSTANCE = new WebChromeClientHostApiCodec();
-
-    private WebChromeClientHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface WebChromeClientHostApi {
-    void create(@NonNull Long instanceId, @NonNull Long webViewClientInstanceId);
+    void create(@NonNull Long instanceId);
 
     /** The codec used by WebChromeClientHostApi. */
     static MessageCodec<Object> getCodec() {
-      return WebChromeClientHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `WebChromeClientHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -2416,19 +2334,12 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
                   }
-                  Number webViewClientInstanceIdArg = (Number) args.get(1);
-                  if (webViewClientInstanceIdArg == null) {
-                    throw new NullPointerException("webViewClientInstanceIdArg unexpectedly null.");
-                  }
-                  api.create(
-                      (instanceIdArg == null) ? null : instanceIdArg.longValue(),
-                      (webViewClientInstanceIdArg == null)
-                          ? null
-                          : webViewClientInstanceIdArg.longValue());
+                  api.create((instanceIdArg == null) ? null : instanceIdArg.longValue());
                   wrapped.put("result", null);
                 } catch (Error | RuntimeException exception) {
                   wrapped.put("error", wrapError(exception));
@@ -2441,14 +2352,6 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class FlutterAssetManagerHostApiCodec extends StandardMessageCodec {
-    public static final FlutterAssetManagerHostApiCodec INSTANCE =
-        new FlutterAssetManagerHostApiCodec();
-
-    private FlutterAssetManagerHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface FlutterAssetManagerHostApi {
     @NonNull
@@ -2459,9 +2362,8 @@ public class GeneratedAndroidWebView {
 
     /** The codec used by FlutterAssetManagerHostApi. */
     static MessageCodec<Object> getCodec() {
-      return FlutterAssetManagerHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `FlutterAssetManagerHostApi` to handle messages through the
      * `binaryMessenger`.
@@ -2477,6 +2379,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   String pathArg = (String) args.get(0);
                   if (pathArg == null) {
                     throw new NullPointerException("pathArg unexpectedly null.");
@@ -2504,6 +2407,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   String nameArg = (String) args.get(0);
                   if (nameArg == null) {
                     throw new NullPointerException("nameArg unexpectedly null.");
@@ -2521,14 +2425,6 @@ public class GeneratedAndroidWebView {
       }
     }
   }
-
-  private static class WebChromeClientFlutterApiCodec extends StandardMessageCodec {
-    public static final WebChromeClientFlutterApiCodec INSTANCE =
-        new WebChromeClientFlutterApiCodec();
-
-    private WebChromeClientFlutterApiCodec() {}
-  }
-
   /** Generated class from Pigeon that represents Flutter messages that can be called from Java. */
   public static class WebChromeClientFlutterApi {
     private final BinaryMessenger binaryMessenger;
@@ -2540,20 +2436,9 @@ public class GeneratedAndroidWebView {
     public interface Reply<T> {
       void reply(T reply);
     }
-
+    /** The codec used by WebChromeClientFlutterApi. */
     static MessageCodec<Object> getCodec() {
-      return WebChromeClientFlutterApiCodec.INSTANCE;
-    }
-
-    public void dispose(@NonNull Long instanceIdArg, Reply<Void> callback) {
-      BasicMessageChannel<Object> channel =
-          new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.WebChromeClientFlutterApi.dispose", getCodec());
-      channel.send(
-          new ArrayList<Object>(Arrays.asList(instanceIdArg)),
-          channelReply -> {
-            callback.reply(null);
-          });
+      return new StandardMessageCodec();
     }
 
     public void onProgressChanged(
@@ -2573,13 +2458,6 @@ public class GeneratedAndroidWebView {
           });
     }
   }
-
-  private static class WebStorageHostApiCodec extends StandardMessageCodec {
-    public static final WebStorageHostApiCodec INSTANCE = new WebStorageHostApiCodec();
-
-    private WebStorageHostApiCodec() {}
-  }
-
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface WebStorageHostApi {
     void create(@NonNull Long instanceId);
@@ -2588,9 +2466,8 @@ public class GeneratedAndroidWebView {
 
     /** The codec used by WebStorageHostApi. */
     static MessageCodec<Object> getCodec() {
-      return WebStorageHostApiCodec.INSTANCE;
+      return new StandardMessageCodec();
     }
-
     /**
      * Sets up an instance of `WebStorageHostApi` to handle messages through the `binaryMessenger`.
      */
@@ -2605,6 +2482,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -2630,6 +2508,7 @@ public class GeneratedAndroidWebView {
                 Map<String, Object> wrapped = new HashMap<>();
                 try {
                   ArrayList<Object> args = (ArrayList<Object>) message;
+                  assert args != null;
                   Number instanceIdArg = (Number) args.get(0);
                   if (instanceIdArg == null) {
                     throw new NullPointerException("instanceIdArg unexpectedly null.");
@@ -2648,7 +2527,8 @@ public class GeneratedAndroidWebView {
     }
   }
 
-  private static Map<String, Object> wrapError(Throwable exception) {
+  @NonNull
+  private static Map<String, Object> wrapError(@NonNull Throwable exception) {
     Map<String, Object> errorMap = new HashMap<>();
     errorMap.put("message", exception.toString());
     errorMap.put("code", exception.getClass().getSimpleName());
