@@ -8,8 +8,6 @@ import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -35,6 +33,8 @@ public class WebViewClientTest {
 
   @Mock public WebView mockWebView;
 
+  @Mock public WebViewClientCompatImpl mockWebViewClient;
+
   InstanceManager instanceManager;
   WebViewClientHostApiImpl hostApiImpl;
   WebViewClientCompatImpl webViewClient;
@@ -48,18 +48,15 @@ public class WebViewClientTest {
     final WebViewClientCreator webViewClientCreator =
         new WebViewClientCreator() {
           @Override
-          public WebViewClient createWebViewClient(
-              WebViewClientFlutterApiImpl flutterApi, boolean shouldOverrideUrlLoading) {
-            webViewClient =
-                (WebViewClientCompatImpl)
-                    super.createWebViewClient(flutterApi, shouldOverrideUrlLoading);
+          public WebViewClient createWebViewClient(WebViewClientFlutterApiImpl flutterApi) {
+            webViewClient = (WebViewClientCompatImpl) super.createWebViewClient(flutterApi);
             return webViewClient;
           }
         };
 
     hostApiImpl =
         new WebViewClientHostApiImpl(instanceManager, webViewClientCreator, mockFlutterApi);
-    hostApiImpl.create(1L, true);
+    hostApiImpl.create(1L);
   }
 
   @After
@@ -72,11 +69,6 @@ public class WebViewClientTest {
     webViewClient.onPageStarted(mockWebView, "https://www.google.com", null);
     verify(mockFlutterApi)
         .onPageStarted(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
-
-    reset(mockFlutterApi);
-    webViewClient.release();
-    webViewClient.onPageStarted(mockWebView, "", null);
-    verify(mockFlutterApi, never()).onPageStarted((WebViewClient) any(), any(), any(), any());
   }
 
   @Test
@@ -90,12 +82,6 @@ public class WebViewClientTest {
             eq("description"),
             eq("https://www.google.com"),
             any());
-
-    reset(mockFlutterApi);
-    webViewClient.release();
-    webViewClient.onReceivedError(mockWebView, 33, "", "");
-    verify(mockFlutterApi, never())
-        .onReceivedError((WebViewClient) any(), any(), any(), any(), any(), any());
   }
 
   @Test
@@ -103,11 +89,6 @@ public class WebViewClientTest {
     webViewClient.shouldOverrideUrlLoading(mockWebView, "https://www.google.com");
     verify(mockFlutterApi)
         .urlLoading(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
-
-    reset(mockFlutterApi);
-    webViewClient.release();
-    webViewClient.shouldOverrideUrlLoading(mockWebView, "");
-    verify(mockFlutterApi, never()).urlLoading((WebViewClient) any(), any(), any(), any());
   }
 
   @Test
@@ -124,5 +105,24 @@ public class WebViewClientTest {
     final GeneratedAndroidWebView.WebResourceRequestData data =
         WebViewClientFlutterApiImpl.createWebResourceRequestData(mockRequest);
     assertEquals(data.getRequestHeaders(), new HashMap<String, String>());
+  }
+
+  @Test
+  public void setReturnValueForShouldOverrideUrlLoading() {
+    final WebViewClientHostApiImpl webViewClientHostApi =
+        new WebViewClientHostApiImpl(
+            instanceManager,
+            new WebViewClientCreator() {
+              @Override
+              public WebViewClient createWebViewClient(WebViewClientFlutterApiImpl flutterApi) {
+                return mockWebViewClient;
+              }
+            },
+            mockFlutterApi);
+
+    instanceManager.addDartCreatedInstance(mockWebViewClient, 0);
+    webViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading(0L, false);
+
+    verify(mockWebViewClient).setReturnValueForShouldOverrideUrlLoading(false);
   }
 }
