@@ -38,7 +38,9 @@ class JavaObject with Copyable {
 
   /// Global instance of [InstanceManager].
   static final InstanceManager globalInstanceManager = InstanceManager(
-    onWeakReferenceRemoved: (_) {},
+    onWeakReferenceRemoved: (int identifier) {
+      JavaObjectHostApiImpl().dispose(identifier);
+    },
   );
 
   /// Pigeon Host Api implementation for [JavaObject].
@@ -76,6 +78,10 @@ class JavaObject with Copyable {
 /// When a [WebView] is no longer needed [release] must be called.
 class WebView extends JavaObject {
   /// Constructs a new WebView.
+  ///
+  /// Due to changes in Flutter 3.0 the [useHybridComposition] doesn't have
+  /// any effect and should not be exposed publicly. More info here:
+  /// https://github.com/flutter/flutter/issues/108106
   WebView({this.useHybridComposition = false}) : super.detached() {
     api.createFromInstance(this);
   }
@@ -89,8 +95,6 @@ class WebView extends JavaObject {
   /// Pigeon Host Api implementation for [WebView].
   @visibleForTesting
   static WebViewHostApiImpl api = WebViewHostApiImpl();
-
-  WebViewClient? _currentWebViewClient;
 
   /// Whether the [WebView] will be rendered with an [AndroidViewSurface].
   ///
@@ -330,8 +334,6 @@ class WebView extends JavaObject {
   ///
   /// This will replace the current handler.
   Future<void> setWebViewClient(WebViewClient webViewClient) {
-    _currentWebViewClient = webViewClient;
-    WebViewClient.api.createFromInstance(webViewClient);
     return api.setWebViewClientFromInstance(this, webViewClient);
   }
 
@@ -375,11 +377,8 @@ class WebView extends JavaObject {
   /// Registers the interface to be used when content can not be handled by the rendering engine, and should be downloaded instead.
   ///
   /// This will replace the current handler.
-  Future<void> setDownloadListener(DownloadListener? listener) async {
-    await Future.wait(<Future<void>>[
-      if (listener != null) DownloadListener.api.createFromInstance(listener),
-      api.setDownloadListenerFromInstance(this, listener)
-    ]);
+  Future<void> setDownloadListener(DownloadListener? listener) {
+    return api.setDownloadListenerFromInstance(this, listener);
   }
 
   /// Sets the chrome handler.
@@ -387,34 +386,13 @@ class WebView extends JavaObject {
   /// This is an implementation of [WebChromeClient] for use in handling
   /// JavaScript dialogs, favicons, titles, and the progress. This will replace
   /// the current handler.
-  Future<void> setWebChromeClient(WebChromeClient? client) async {
-    // WebView requires a WebViewClient because of a bug fix that makes
-    // calls to WebViewClient.requestLoading/WebViewClient.urlLoading when a new
-    // window is opened. This is to make sure a url opened by `Window.open` has
-    // a secure url.
-    assert(
-      _currentWebViewClient != null,
-      "Can't set a WebChromeClient without setting a WebViewClient first.",
-    );
-    await Future.wait(<Future<void>>[
-      if (client != null)
-        WebChromeClient.api.createFromInstance(client, _currentWebViewClient!),
-      api.setWebChromeClientFromInstance(this, client),
-    ]);
+  Future<void> setWebChromeClient(WebChromeClient? client) {
+    return api.setWebChromeClientFromInstance(this, client);
   }
 
   /// Sets the background color of this WebView.
   Future<void> setBackgroundColor(Color color) {
     return api.setBackgroundColorFromInstance(this, color.value);
-  }
-
-  /// Releases all resources used by the [WebView].
-  ///
-  /// Any methods called after [release] will throw an exception.
-  Future<void> release() {
-    _currentWebViewClient = null;
-    WebSettings.api.disposeFromInstance(settings);
-    return api.disposeFromInstance(this);
   }
 
   @override
@@ -624,9 +602,10 @@ class JavaScriptChannel extends JavaObject {
   /// Constructs a [JavaScriptChannel].
   JavaScriptChannel(
     this.channelName, {
-    void Function(String message)? postMessage,
+    required this.postMessage,
   }) : super.detached() {
     AndroidWebViewFlutterApis.instance.ensureSetUp();
+    api.createFromInstance(this);
   }
 
   /// Constructs a [JavaScriptChannel] without creating the associated Java
@@ -636,7 +615,7 @@ class JavaScriptChannel extends JavaObject {
   /// create copies.
   JavaScriptChannel.detached(
     this.channelName, {
-    void Function(String message)? postMessage,
+    required this.postMessage,
   }) : super.detached();
 
   /// Pigeon Host Api implementation for [JavaScriptChannel].
@@ -647,7 +626,7 @@ class JavaScriptChannel extends JavaObject {
   final String channelName;
 
   /// Callback method when javaScript calls `postMessage` on the object instance passed.
-  void postMessage(String message) {}
+  final void Function(String message) postMessage;
 
   @override
   JavaScriptChannel copy() {
@@ -659,26 +638,15 @@ class JavaScriptChannel extends JavaObject {
 class WebViewClient extends JavaObject {
   /// Constructs a [WebViewClient].
   WebViewClient({
-    this.shouldOverrideUrlLoading = true,
-    void Function(WebView webView, String url)? onPageStarted,
-    void Function(WebView webView, String url)? onPageFinished,
-    void Function(
-      WebView webView,
-      WebResourceRequest request,
-      WebResourceError error,
-    )?
-        onReceivedRequestError,
-    void Function(
-      WebView webView,
-      int errorCode,
-      String description,
-      String failingUrl,
-    )?
-        onReceivedError,
-    void Function(WebView webView, WebResourceRequest request)? requestLoading,
-    void Function(WebView webView, String url)? urlLoading,
+    this.onPageStarted,
+    this.onPageFinished,
+    this.onReceivedRequestError,
+    @Deprecated('Only called on Android version < 23.') this.onReceivedError,
+    this.requestLoading,
+    this.urlLoading,
   }) : super.detached() {
     AndroidWebViewFlutterApis.instance.ensureSetUp();
+    api.createFromInstance(this);
   }
 
   /// Constructs a [WebViewClient] without creating the associated Java object.
@@ -686,24 +654,12 @@ class WebViewClient extends JavaObject {
   /// This should only be used by subclasses created by this library or to
   /// create copies.
   WebViewClient.detached({
-    this.shouldOverrideUrlLoading = true,
-    void Function(WebView webView, String url)? onPageStarted,
-    void Function(WebView webView, String url)? onPageFinished,
-    void Function(
-      WebView webView,
-      WebResourceRequest request,
-      WebResourceError error,
-    )?
-        onReceivedRequestError,
-    void Function(
-      WebView webView,
-      int errorCode,
-      String description,
-      String failingUrl,
-    )?
-        onReceivedError,
-    void Function(WebView webView, WebResourceRequest request)? requestLoading,
-    void Function(WebView webView, String url)? urlLoading,
+    this.onPageStarted,
+    this.onPageFinished,
+    this.onReceivedRequestError,
+    @Deprecated('Only called on Android version < 23.') this.onReceivedError,
+    this.requestLoading,
+    this.urlLoading,
   }) : super.detached();
 
   /// User authentication failed on server.
@@ -790,20 +746,6 @@ class WebViewClient extends JavaObject {
   @visibleForTesting
   static WebViewClientHostApiImpl api = WebViewClientHostApiImpl();
 
-  /// Whether loading a url should be overridden.
-  ///
-  /// In Java, `shouldOverrideUrlLoading()` and `shouldOverrideRequestLoading()`
-  /// callbacks must synchronously return a boolean. This sets the default
-  /// return value.
-  ///
-  /// Setting [shouldOverrideUrlLoading] to true causes the current [WebView] to
-  /// abort loading the URL, while returning false causes the [WebView] to
-  /// continue loading the URL as usual. [requestLoading] or [urlLoading] will
-  /// still be called either way.
-  ///
-  /// Defaults to true.
-  final bool shouldOverrideUrlLoading;
-
   /// Notify the host application that a page has started loading.
   ///
   /// This method is called once for each main frame load so a page with iframes
@@ -812,7 +754,7 @@ class WebViewClient extends JavaObject {
   /// embedded frame changes, i.e. clicking a link whose target is an iframe, it
   /// will also not be called for fragment navigations (navigations to
   /// #fragment_id).
-  void onPageStarted(WebView webView, String url) {}
+  final void Function(WebView webView, String url)? onPageStarted;
 
   // TODO(bparrishMines): Update documentation when WebView.postVisualStateCallback is added.
   /// Notify the host application that a page has finished loading.
@@ -820,7 +762,7 @@ class WebViewClient extends JavaObject {
   /// This method is called only for main frame. Receiving an [onPageFinished]
   /// callback does not guarantee that the next frame drawn by WebView will
   /// reflect the state of the DOM at this point.
-  void onPageFinished(WebView webView, String url) {}
+  final void Function(WebView webView, String url)? onPageFinished;
 
   /// Report web resource loading error to the host application.
   ///
@@ -829,48 +771,58 @@ class WebViewClient extends JavaObject {
   /// be called for any resource (iframe, image, etc.), not just for the main
   /// page. Thus, it is recommended to perform minimum required work in this
   /// callback.
-  void onReceivedRequestError(
+  final void Function(
     WebView webView,
     WebResourceRequest request,
     WebResourceError error,
-  ) {}
+  )? onReceivedRequestError;
 
   /// Report an error to the host application.
   ///
   /// These errors are unrecoverable (i.e. the main resource is unavailable).
   /// The errorCode parameter corresponds to one of the error* constants.
   @Deprecated('Only called on Android version < 23.')
-  void onReceivedError(
+  final void Function(
     WebView webView,
     int errorCode,
     String description,
     String failingUrl,
-  ) {}
+  )? onReceivedError;
 
-  // TODO(bparrishMines): Update documentation once synchronous url handling is supported.
-  /// When a URL is about to be loaded in the current [WebView].
+  /// When the current [WebView] wants to load a URL.
   ///
-  /// If a [WebViewClient] is not provided, by default [WebView] will ask
-  /// Activity Manager to choose the proper handler for the URL. If a
-  /// [WebViewClient] is provided, setting [shouldOverrideUrlLoading] to true
-  /// causes the current [WebView] to abort loading the URL, while returning
-  /// false causes the [WebView] to continue loading the URL as usual.
-  void requestLoading(WebView webView, WebResourceRequest request) {}
+  /// The value set by [setSynchronousReturnValueForShouldOverrideUrlLoading]
+  /// indicates whether the [WebView] loaded the request.
+  final void Function(WebView webView, WebResourceRequest request)?
+      requestLoading;
 
-  // TODO(bparrishMines): Update documentation once synchronous url handling is supported.
-  /// When a URL is about to be loaded in the current [WebView].
+  /// When the current [WebView] wants to load a URL.
   ///
-  /// If a [WebViewClient] is not provided, by default [WebView] will ask
-  /// Activity Manager to choose the proper handler for the URL. If a
-  /// [WebViewClient] is provided, setting [shouldOverrideUrlLoading] to true
-  /// causes the current [WebView] to abort loading the URL, while returning
-  /// false causes the [WebView] to continue loading the URL as usual.
-  void urlLoading(WebView webView, String url) {}
+  /// The value set by [setSynchronousReturnValueForShouldOverrideUrlLoading]
+  /// indicates whether the [WebView] loaded the URL.
+  final void Function(WebView webView, String url)? urlLoading;
+
+  /// Sets the required synchronous return value for the Java method,
+  /// `WebViewClient.shouldOverrideUrlLoading(...)`.
+  ///
+  /// The Java method, `WebViewClient.shouldOverrideUrlLoading(...)`, requires
+  /// a boolean to be returned and this method sets the returned value for all
+  /// calls to the Java method.
+  ///
+  /// Setting this to true causes the current [WebView] to abort loading any URL
+  /// received by [requestLoading] or [urlLoading], while setting this to false
+  /// causes the [WebView] to continue loading a URL as usual.
+  ///
+  /// Defaults to false.
+  Future<void> setSynchronousReturnValueForShouldOverrideUrlLoading(
+    bool value,
+  ) {
+    return api.setShouldOverrideUrlLoadingReturnValueFromInstance(this, value);
+  }
 
   @override
   WebViewClient copy() {
     return WebViewClient.detached(
-      shouldOverrideUrlLoading: shouldOverrideUrlLoading,
       onPageStarted: onPageStarted,
       onPageFinished: onPageFinished,
       onReceivedRequestError: onReceivedRequestError,
@@ -885,17 +837,9 @@ class WebViewClient extends JavaObject {
 /// engine for [WebView], and should be downloaded instead.
 class DownloadListener extends JavaObject {
   /// Constructs a [DownloadListener].
-  DownloadListener({
-    void Function(
-      String url,
-      String userAgent,
-      String contentDisposition,
-      String mimetype,
-      int contentLength,
-    )?
-        onDownloadStart,
-  }) : super.detached() {
+  DownloadListener({required this.onDownloadStart}) : super.detached() {
     AndroidWebViewFlutterApis.instance.ensureSetUp();
+    api.createFromInstance(this);
   }
 
   /// Constructs a [DownloadListener] without creating the associated Java
@@ -903,43 +847,33 @@ class DownloadListener extends JavaObject {
   ///
   /// This should only be used by subclasses created by this library or to
   /// create copies.
-  DownloadListener.detached({
-    void Function(
-      String url,
-      String userAgent,
-      String contentDisposition,
-      String mimetype,
-      int contentLength,
-    )?
-        onDownloadStart,
-  }) : super.detached();
+  DownloadListener.detached({required this.onDownloadStart}) : super.detached();
 
   /// Pigeon Host Api implementation for [DownloadListener].
   @visibleForTesting
   static DownloadListenerHostApiImpl api = DownloadListenerHostApiImpl();
 
   /// Notify the host application that a file should be downloaded.
-  void onDownloadStart(
+  final void Function(
     String url,
     String userAgent,
     String contentDisposition,
     String mimetype,
     int contentLength,
-  ) {}
+  ) onDownloadStart;
 
   @override
   DownloadListener copy() {
-    return DownloadListener(onDownloadStart: onDownloadStart);
+    return DownloadListener.detached(onDownloadStart: onDownloadStart);
   }
 }
 
 /// Handles JavaScript dialogs, favicons, titles, and the progress for [WebView].
 class WebChromeClient extends JavaObject {
   /// Constructs a [WebChromeClient].
-  WebChromeClient({
-    void Function(WebView webView, int progress)? onProgressChanged,
-  }) : super.detached() {
+  WebChromeClient({this.onProgressChanged}) : super.detached() {
     AndroidWebViewFlutterApis.instance.ensureSetUp();
+    api.createFromInstance(this);
   }
 
   /// Constructs a [WebChromeClient] without creating the associated Java
@@ -947,16 +881,14 @@ class WebChromeClient extends JavaObject {
   ///
   /// This should only be used by subclasses created by this library or to
   /// create copies.
-  WebChromeClient.detached({
-    void Function(WebView webView, int progress)? onProgressChanged,
-  }) : super.detached();
+  WebChromeClient.detached({this.onProgressChanged}) : super.detached();
 
   /// Pigeon Host Api implementation for [WebChromeClient].
   @visibleForTesting
   static WebChromeClientHostApiImpl api = WebChromeClientHostApiImpl();
 
   /// Notify the host application that a file should be downloaded.
-  void onProgressChanged(WebView webView, int progress) {}
+  final void Function(WebView webView, int progress)? onProgressChanged;
 
   @override
   WebChromeClient copy() {
