@@ -88,8 +88,6 @@ abstract class CookieManagerHostApi {
 abstract class WebViewHostApi {
   void create(int instanceId, bool useHybridComposition);
 
-  void dispose(int instanceId);
-
   void loadData(
     int instanceId,
     String data,
@@ -169,8 +167,6 @@ abstract class WebViewHostApi {
 abstract class WebSettingsHostApi {
   void create(int instanceId, int webViewInstanceId);
 
-  void dispose(int instanceId);
-
   void setDomStorageEnabled(int instanceId, bool flag);
 
   void setJavaScriptCanOpenWindowsAutomatically(int instanceId, bool flag);
@@ -203,20 +199,21 @@ abstract class JavaScriptChannelHostApi {
 
 @FlutterApi()
 abstract class JavaScriptChannelFlutterApi {
-  void dispose(int instanceId);
-
   void postMessage(int instanceId, String message);
 }
 
 @HostApi(dartHostTestHandler: 'TestWebViewClientHostApi')
 abstract class WebViewClientHostApi {
-  void create(int instanceId, bool shouldOverrideUrlLoading);
+  void create(int instanceId);
+
+  void setSynchronousReturnValueForShouldOverrideUrlLoading(
+    int instanceId,
+    bool value,
+  );
 }
 
 @FlutterApi()
 abstract class WebViewClientFlutterApi {
-  void dispose(int instanceId);
-
   void onPageStarted(int instanceId, int webViewInstanceId, String url);
 
   void onPageFinished(int instanceId, int webViewInstanceId, String url);
@@ -252,8 +249,6 @@ abstract class DownloadListenerHostApi {
 
 @FlutterApi()
 abstract class DownloadListenerFlutterApi {
-  void dispose(int instanceId);
-
   void onDownloadStart(
     int instanceId,
     String url,
@@ -266,7 +261,7 @@ abstract class DownloadListenerFlutterApi {
 
 @HostApi(dartHostTestHandler: 'TestWebChromeClientHostApi')
 abstract class WebChromeClientHostApi {
-  void create(int instanceId, int webViewClientInstanceId);
+  void create(int instanceId);
 }
 
 @HostApi(dartHostTestHandler: 'TestAssetManagerHostApi')
@@ -278,8 +273,6 @@ abstract class FlutterAssetManagerHostApi {
 
 @FlutterApi()
 abstract class WebChromeClientFlutterApi {
-  void dispose(int instanceId);
-
   void onProgressChanged(int instanceId, int webViewInstanceId, int progress);
 }
 
