@@ -1,7 +1,8 @@
-// Mocks generated by Mockito 5.2.0 from annotations
-// in webview_flutter_android/test/webview_android_cookie_manager_test.dart.
+// Mocks generated by Mockito 5.3.2 from annotations
+// in webview_flutter_android/test/android_webview_cookie_manager_test.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i3;
 
 import 'package:mockito/mockito.dart' as _i1;
@@ -16,6 +17,7 @@ import 'package:webview_flutter_android/src/android_webview.dart' as _i2;
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
 /// A class which mocks [CookieManager].
 ///
@@ -26,12 +28,27 @@ class MockCookieManager extends _i1.Mock implements _i2.CookieManager {
   }
 
   @override
-  _i3.Future<void> setCookie(String? url, String? value) =>
-      (super.noSuchMethod(Invocation.method(#setCookie, [url, value]),
-          returnValue: Future<void>.value(),
-          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
+  _i3.Future<void> setCookie(
+    String? url,
+    String? value,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setCookie,
+          [
+            url,
+            value,
+          ],
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
   @override
-  _i3.Future<bool> clearCookies() =>
-      (super.noSuchMethod(Invocation.method(#clearCookies, []),
-          returnValue: Future<bool>.value(false)) as _i3.Future<bool>);
+  _i3.Future<bool> clearCookies() => (super.noSuchMethod(
+        Invocation.method(
+          #clearCookies,
+          [],
+        ),
+        returnValue: _i3.Future<bool>.value(false),
+      ) as _i3.Future<bool>);
 }
