@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v4.0.2), do not edit directly.
+// Autogenerated from Pigeon (v4.2.3), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 // ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, unnecessary_import
 // ignore_for_file: avoid_relative_lib_imports
@@ -13,12 +13,13 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:webview_flutter_android/src/android_webview.pigeon.dart';
 
-class _TestJavaObjectHostApiCodec extends StandardMessageCodec {
-  const _TestJavaObjectHostApiCodec();
-}
-
+/// Handles methods calls to the native Java Object class.
+///
+/// Also handles calls to remove the reference to an instance with `dispose`.
+///
+/// See https://docs.oracle.com/javase/7/docs/api/java/lang/Object.html.
 abstract class TestJavaObjectHostApi {
-  static const MessageCodec<Object?> codec = _TestJavaObjectHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   void dispose(int identifier);
   static void setup(TestJavaObjectHostApi? api,
@@ -73,7 +74,6 @@ abstract class TestWebViewHostApi {
   static const MessageCodec<Object?> codec = _TestWebViewHostApiCodec();
 
   void create(int instanceId, bool useHybridComposition);
-  void dispose(int instanceId);
   void loadData(
       int instanceId, String data, String? mimeType, String? encoding);
   void loadDataWithBaseUrl(int instanceId, String? baseUrl, String data,
@@ -121,25 +121,6 @@ abstract class TestWebViewHostApi {
           assert(arg_useHybridComposition != null,
               'Argument for dev.flutter.pigeon.WebViewHostApi.create was null, expected non-null bool.');
           api.create(arg_instanceId!, arg_useHybridComposition!);
-          return <Object?, Object?>{};
-        });
-      }
-    }
-    {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.WebViewHostApi.dispose', codec,
-          binaryMessenger: binaryMessenger);
-      if (api == null) {
-        channel.setMockMessageHandler(null);
-      } else {
-        channel.setMockMessageHandler((Object? message) async {
-          assert(message != null,
-              'Argument for dev.flutter.pigeon.WebViewHostApi.dispose was null.');
-          final List<Object?> args = (message as List<Object?>?)!;
-          final int? arg_instanceId = (args[0] as int?);
-          assert(arg_instanceId != null,
-              'Argument for dev.flutter.pigeon.WebViewHostApi.dispose was null, expected non-null int.');
-          api.dispose(arg_instanceId!);
           return <Object?, Object?>{};
         });
       }
@@ -684,15 +665,10 @@ abstract class TestWebViewHostApi {
   }
 }
 
-class _TestWebSettingsHostApiCodec extends StandardMessageCodec {
-  const _TestWebSettingsHostApiCodec();
-}
-
 abstract class TestWebSettingsHostApi {
-  static const MessageCodec<Object?> codec = _TestWebSettingsHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   void create(int instanceId, int webViewInstanceId);
-  void dispose(int instanceId);
   void setDomStorageEnabled(int instanceId, bool flag);
   void setJavaScriptCanOpenWindowsAutomatically(int instanceId, bool flag);
   void setSupportMultipleWindows(int instanceId, bool support);
@@ -725,25 +701,6 @@ abstract class TestWebSettingsHostApi {
           assert(arg_webViewInstanceId != null,
               'Argument for dev.flutter.pigeon.WebSettingsHostApi.create was null, expected non-null int.');
           api.create(arg_instanceId!, arg_webViewInstanceId!);
-          return <Object?, Object?>{};
-        });
-      }
-    }
-    {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.WebSettingsHostApi.dispose', codec,
-          binaryMessenger: binaryMessenger);
-      if (api == null) {
-        channel.setMockMessageHandler(null);
-      } else {
-        channel.setMockMessageHandler((Object? message) async {
-          assert(message != null,
-              'Argument for dev.flutter.pigeon.WebSettingsHostApi.dispose was null.');
-          final List<Object?> args = (message as List<Object?>?)!;
-          final int? arg_instanceId = (args[0] as int?);
-          assert(arg_instanceId != null,
-              'Argument for dev.flutter.pigeon.WebSettingsHostApi.dispose was null, expected non-null int.');
-          api.dispose(arg_instanceId!);
           return <Object?, Object?>{};
         });
       }
@@ -1019,13 +976,8 @@ abstract class TestWebSettingsHostApi {
   }
 }
 
-class _TestJavaScriptChannelHostApiCodec extends StandardMessageCodec {
-  const _TestJavaScriptChannelHostApiCodec();
-}
-
 abstract class TestJavaScriptChannelHostApi {
-  static const MessageCodec<Object?> codec =
-      _TestJavaScriptChannelHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   void create(int instanceId, String channelName);
   static void setup(TestJavaScriptChannelHostApi? api,
@@ -1055,14 +1007,12 @@ abstract class TestJavaScriptChannelHostApi {
   }
 }
 
-class _TestWebViewClientHostApiCodec extends StandardMessageCodec {
-  const _TestWebViewClientHostApiCodec();
-}
-
 abstract class TestWebViewClientHostApi {
-  static const MessageCodec<Object?> codec = _TestWebViewClientHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
-  void create(int instanceId, bool shouldOverrideUrlLoading);
+  void create(int instanceId);
+  void setSynchronousReturnValueForShouldOverrideUrlLoading(
+      int instanceId, bool value);
   static void setup(TestWebViewClientHostApi? api,
       {BinaryMessenger? binaryMessenger}) {
     {
@@ -1079,10 +1029,31 @@ abstract class TestWebViewClientHostApi {
           final int? arg_instanceId = (args[0] as int?);
           assert(arg_instanceId != null,
               'Argument for dev.flutter.pigeon.WebViewClientHostApi.create was null, expected non-null int.');
-          final bool? arg_shouldOverrideUrlLoading = (args[1] as bool?);
-          assert(arg_shouldOverrideUrlLoading != null,
-              'Argument for dev.flutter.pigeon.WebViewClientHostApi.create was null, expected non-null bool.');
-          api.create(arg_instanceId!, arg_shouldOverrideUrlLoading!);
+          api.create(arg_instanceId!);
+          return <Object?, Object?>{};
+        });
+      }
+    }
+    {
+      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.WebViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading',
+          codec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        channel.setMockMessageHandler(null);
+      } else {
+        channel.setMockMessageHandler((Object? message) async {
+          assert(message != null,
+              'Argument for dev.flutter.pigeon.WebViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final int? arg_instanceId = (args[0] as int?);
+          assert(arg_instanceId != null,
+              'Argument for dev.flutter.pigeon.WebViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading was null, expected non-null int.');
+          final bool? arg_value = (args[1] as bool?);
+          assert(arg_value != null,
+              'Argument for dev.flutter.pigeon.WebViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading was null, expected non-null bool.');
+          api.setSynchronousReturnValueForShouldOverrideUrlLoading(
+              arg_instanceId!, arg_value!);
           return <Object?, Object?>{};
         });
       }
@@ -1090,13 +1061,8 @@ abstract class TestWebViewClientHostApi {
   }
 }
 
-class _TestDownloadListenerHostApiCodec extends StandardMessageCodec {
-  const _TestDownloadListenerHostApiCodec();
-}
-
 abstract class TestDownloadListenerHostApi {
-  static const MessageCodec<Object?> codec =
-      _TestDownloadListenerHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   void create(int instanceId);
   static void setup(TestDownloadListenerHostApi? api,
@@ -1123,14 +1089,10 @@ abstract class TestDownloadListenerHostApi {
   }
 }
 
-class _TestWebChromeClientHostApiCodec extends StandardMessageCodec {
-  const _TestWebChromeClientHostApiCodec();
-}
-
 abstract class TestWebChromeClientHostApi {
-  static const MessageCodec<Object?> codec = _TestWebChromeClientHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
-  void create(int instanceId, int webViewClientInstanceId);
+  void create(int instanceId);
   static void setup(TestWebChromeClientHostApi? api,
       {BinaryMessenger? binaryMessenger}) {
     {
@@ -1147,10 +1109,7 @@ abstract class TestWebChromeClientHostApi {
           final int? arg_instanceId = (args[0] as int?);
           assert(arg_instanceId != null,
               'Argument for dev.flutter.pigeon.WebChromeClientHostApi.create was null, expected non-null int.');
-          final int? arg_webViewClientInstanceId = (args[1] as int?);
-          assert(arg_webViewClientInstanceId != null,
-              'Argument for dev.flutter.pigeon.WebChromeClientHostApi.create was null, expected non-null int.');
-          api.create(arg_instanceId!, arg_webViewClientInstanceId!);
+          api.create(arg_instanceId!);
           return <Object?, Object?>{};
         });
       }
@@ -1158,12 +1117,8 @@ abstract class TestWebChromeClientHostApi {
   }
 }
 
-class _TestAssetManagerHostApiCodec extends StandardMessageCodec {
-  const _TestAssetManagerHostApiCodec();
-}
-
 abstract class TestAssetManagerHostApi {
-  static const MessageCodec<Object?> codec = _TestAssetManagerHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   List<String?> list(String path);
   String getAssetFilePathByName(String name);
@@ -1211,12 +1166,8 @@ abstract class TestAssetManagerHostApi {
   }
 }
 
-class _TestWebStorageHostApiCodec extends StandardMessageCodec {
-  const _TestWebStorageHostApiCodec();
-}
-
 abstract class TestWebStorageHostApi {
-  static const MessageCodec<Object?> codec = _TestWebStorageHostApiCodec();
+  static const MessageCodec<Object?> codec = StandardMessageCodec();
 
   void create(int instanceId);
   void deleteAllData(int instanceId);
